refactor: move pool-increase components
