+++ packages/web/src/components/decrease/decrease-liquidity/DecreaseLiquidity.tsx
@@ -3,12 +3,15 @@ import { useTranslation } from "react-i18next";
 
 import Button, { ButtonHierarchy } from "@components/common/button/Button";
 import { RANGE_STATUS_OPTION } from "@constants/option.constant";
-import { IPooledTokenInfo } from "@hooks/decrease/use-decrease-handle";
-import { IPriceRange } from "@hooks/increase/use-increase-handle";
+import {
+  IPooledTokenInfo,
+  IPriceRange,
+} from "@hooks/decrease/use-decrease-handle";
 import { TokenModel } from "@models/token/token-model";
 
 import DecreaseAmountPosition from "../decrease-select-position/DecreaseAmount";
 import DecreaseSelectPosition from "../decrease-select-position/DecreaseSelectPosition";
+
 import { DecreaseLiquidityWrapper } from "./DecreaseLiquidity.styles";
 
 interface DecreaseLiquidityProps {
