+++ packages/web/src/components/decrease/decrease-select-position/DecreaseSelectPosition.tsx
@@ -1,14 +1,16 @@
+import React from "react";
+import { useTranslation } from "react-i18next";
+
 import Badge, { BADGE_TYPE } from "@components/common/badge/Badge";
 import DoubleLogo from "@components/common/double-logo/DoubleLogo";
 import MissingLogo from "@components/common/missing-logo/MissingLogo";
 import RangeBadge from "@components/common/range-badge/RangeBadge";
 import { RANGE_STATUS_OPTION } from "@constants/option.constant";
 import { useWindowSize } from "@hooks/common/use-window-size";
-import { IPriceRange } from "@hooks/increase/use-increase-handle";
+import { IPriceRange } from "@hooks/decrease/use-decrease-handle";
 import { TokenModel } from "@models/token/token-model";
 import { DEVICE_TYPE } from "@styles/media";
-import React from "react";
-import { useTranslation } from "react-i18next";
+
 import { DecreaseSelectPositionWrapper } from "./DecreaseSelectPosition.styles";
 
 export interface DecreaseSelectPositionProps {
