+++ packages/web/src/views/pool/common/components/increase-max-min/IncreaseMaxMin.styles.ts
