+++ packages/web/src/views/pool/pool-increase-liquidity/PoolIncreaseLiquidity.tsx
@@ -5,7 +5,6 @@ import Footer from "@components/common/footer/Footer";
 import { PAGE_PATH, QUERY_PARAMETER } from "@constants/page.constant";
 import BreadcrumbsContainer from "@containers/breadcrumbs-container/BreadcrumbsContainer";
 import HeaderContainer from "@containers/header-container/HeaderContainer";
-import IncreaseLiquidityContainer from "@containers/increase-liquidity-container/IncreaseLiquidityContainer";
 import useRouter from "@hooks/common/use-custom-router";
 import { useLoading } from "@hooks/common/use-loading";
 import { useWindowSize } from "@hooks/common/use-window-size";
@@ -15,6 +14,7 @@ import { makeRouteUrl } from "@utils/page.utils";
 import { useGetPoolDetailByPath } from "src/react-query/pools";
 
 import IncreaseLiquidityLayout from "./IncreaseLiquidityLayout";
+import IncreaseLiquidityContainer from "./containers/increase-liquidity-container/IncreaseLiquidityContainer";
 
 const PoolIncreaseLiquidity: React.FC = () => {
   const { t } = useTranslation();
