+++ packages/web/src/views/pool/pool-increase-liquidity/components/increase-position-modal/increase-amount-info/IncreaseAmountInfo.styles.ts
